split functions into modular peices and implement reset logic
