+++ src/index.js
@@ -3,4 +3,3 @@ import "./css/reset.css";
 import "./css/style.css";
 
 createCharacters();
-selectOneCharacter();
