+++ src/javascript/domManipulation/clearArena.js
@@ -1,0 +1,4 @@
+export let clearArena = () => {
+  let arena = document.getElementById("arena");
+  arena.innerHTML = "";
+};
