+++ src/javascript/domManipulation/clearHealthBars.js
@@ -1,0 +1,4 @@
+export let clearHealthBars = () => {
+  let healthBars = document.getElementById("health-bars-container");
+  healthBars.innerHTML = "";
+};
