--- src/javascript/domManipulation/selectCharacter.js
@@ -1,1 +1,0 @@
-export const selectCharacter = () => {};
