+++ src/javascript/gameLogic/resetGame.js
@@ -1,0 +1,15 @@
+import { renderOneCharacter } from "../render/renderOneCharacter";
+import { characters } from "../characterConstructor";
+import { clearArena } from "../domManipulation/clearArena";
+import { clearHealthBars } from "../domManipulation/clearHealthBars";
+
+export let resetGame = () => {
+  let allCharacters = document.getElementById("all-characters");
+  allCharacters.innerHTML = "";
+  allCharacters.style.display = "flex";
+  for (let i = 0; i < characters.length; i++) {
+    renderOneCharacter(characters[i], allCharacters);
+  }
+  clearArena();
+  clearHealthBars();
+};
