+++ src/javascript/render/renderAllCharacters.js
@@ -1,0 +1,8 @@
+import { renderOneCharacter } from "./renderOneCharacter";
+
+export const renderAllCharacters = (characters) => {
+  let allCharacters = document.getElementById("all-characters");
+  for (let i = 0; i < characters.length; i++) {
+    renderOneCharacter(characters[i], allCharacters);
+  }
+};
