+++ src/javascript/state/index.js
@@ -1,4 +1,7 @@
-export let globalState = {
+export let state = {
   user: {},
   opponent: {},
+  characterSelected: false,
+  selected: 0,
+  opponentSelected: false,
 };
